more validations before trying to delete products and branches

+++ src/Ambev.DeveloperEvaluation.Application/Products/DeleteProduct/DeleteProductHandler.cs
@@ -1,4 +1,6 @@
-﻿using Ambev.DeveloperEvaluation.Domain.Repositories;
+﻿using Ambev.DeveloperEvaluation.Application.Services;
+using Ambev.DeveloperEvaluation.Domain.Entities;
+using Ambev.DeveloperEvaluation.Domain.Repositories;
 using FluentValidation;
 using MediatR;
 
@@ -7,10 +9,12 @@ namespace Ambev.DeveloperEvaluation.Application.Products.DeleteProduct
     public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, DeleteProductResult>
     {
         private readonly IProductRepository _productRepository;
+        private readonly IQueryDatabaseService _queryDbService;
 
-        public DeleteProductHandler(IProductRepository productRepository)
+        public DeleteProductHandler(IProductRepository productRepository, IQueryDatabaseService queryDbService)
         {
             _productRepository = productRepository;
+            _queryDbService = queryDbService;
         }
 
         public async Task<DeleteProductResult> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
@@ -21,9 +25,17 @@ namespace Ambev.DeveloperEvaluation.Application.Products.DeleteProduct
             if (!validationResult.IsValid)
                 throw new ValidationException(validationResult.Errors);
 
+            var existing = await _productRepository.GetByIdAsync(command.Id, cancellationToken);
+            if (existing == null)
+                throw new KeyNotFoundException($"Product with ID {command.Id} not found");
+
+            var hasProductInItems = await _queryDbService.ProductsInItem(command.Id.ToString());
+            if (hasProductInItems)
+                throw new InvalidOperationException($"Product with ID {command.Id} cannot be deleted because it is referenced in sales items.");
+
             var success = await _productRepository.DeleteAsync(command.Id, cancellationToken);
             if (!success)
-                throw new KeyNotFoundException($"Record with ID {command.Id} not found");
+                throw new KeyNotFoundException($"Product with ID {command.Id} not found");
 
             return new DeleteProductResult { Success = true };
         }
